requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
reportlab==4.0.7
python-nmap==0.7.1
gunicorn==21.2.0
urllib3==2.1.0
yara-python==4.5.1
